+++ c/csv/fread.h
@@ -17,7 +17,6 @@
 extern const long double pow10lookup[701];
 extern const uint8_t hexdigits[256];
 extern const uint8_t allowedseps[128];
-const char* strlim(const char* ch, size_t limit);
 
 #define JUMPLINES 100    // at each of the 100 jumps how many lines to guess column types (10,000 sample lines)