+++ c/options.h
@@ -20,6 +20,7 @@ extern size_t sort_max_chunk_length;
 extern int8_t sort_max_radix_bits;
 extern int8_t sort_over_radix_bits;
 extern int32_t sort_nthreads;
+extern bool fread_anonymize;
 
 void set_nthreads(int32_t n);
 void set_core_logger(PyObject*);
@@ -29,6 +30,7 @@ void set_sort_max_chunk_length(int64_t n);
 void set_sort_max_radix_bits(int64_t n);
 void set_sort_over_radix_bits(int64_t n);
 void set_sort_nthreads(int32_t n);
+void set_fread_anonymize(int8_t v);
 
 
 DECLARE_FUNCTION(